finished header component
